move Route for login and register into app container
