+++ src/components/HeaderLogin/HeaderLogin.jsx
@@ -1,18 +1,8 @@
 import React from 'react';
-import { Route, Link } from 'react-router-dom';
-
-const Login = () => {
-  return <div>Login Page</div>;
-};
-
-const Register = () => {
-  return <div>Register Page</div>;
-};
+import { Link } from 'react-router-dom';
 
 const HeaderLogin = props => {
   const { currentUser, logout } = props;
-
-  console.log(currentUser);
 
   if (!currentUser) {
     return (
@@ -24,9 +14,6 @@ const HeaderLogin = props => {
         <Link to="/register">
           <button>Register</button>
         </Link>
-
-        <Route path="/login" component={Login} />
-        <Route path="/register" component={Register} />
       </div>
     );
   } else {
